+++ src/components/card-list/card/style.css
@@ -1,4 +1,4 @@
-@import '../../color.css';
+@import '../../../color.css';
 
 .card {
   position: relative;
