+++ src/components/card-list/style.css
@@ -1,4 +1,4 @@
-@import '../color.css';
+@import '../../color.css';
 
 .folder-menu {
   width: 1060px;
