+++ src/components/favorites/index.js
