+++ src/components/favorites/style.css
