+++ src/components/header/style.css
@@ -1,5 +1,5 @@
-@import '../color.css';
-@import '../reset.css';
+@import '../../color.css';
+@import '../../reset.css';
 
 .header {
   width: 100%;
