+++ src/components/search/style.css
@@ -1,4 +1,4 @@
-@import '../color.css';
+@import '../../color.css';
 
 .search__form {
   width: 1060px;
